Use nested models in Gradle instead of POJOs
