+++ micronaut-build-plugin-sourcegen-generator/src/main/java/io/micronaut/sourcegen/generator/visitors/gradle/GradlePluginUtils.java
@@ -27,8 +27,7 @@ import io.micronaut.sourcegen.annotations.GenerateGradlePlugin.GenerateGradleTas
 import io.micronaut.sourcegen.annotations.GenerateGradlePlugin.Type;
 import io.micronaut.sourcegen.generator.visitors.JavadocUtils;
 import io.micronaut.sourcegen.generator.visitors.JavadocUtils.TypeJavadoc;
-import io.micronaut.sourcegen.generator.visitors.ModelUtils;
-import io.micronaut.sourcegen.generator.visitors.ModelUtils.GeneratedModel;
+import io.micronaut.sourcegen.generator.visitors.ModelBuilder;
 import io.micronaut.sourcegen.generator.visitors.PluginUtils;
 import io.micronaut.sourcegen.generator.visitors.PluginUtils.ParameterConfig;
 import io.micronaut.sourcegen.model.TypeDef;
@@ -87,12 +86,12 @@ public final class GradlePluginUtils {
                 + annotation.stringValue("source"));
         }
 
-        List<GeneratedModel> generatedModels = new ArrayList<>();
         TypeJavadoc javadoc = JavadocUtils.getTaskJavadoc(context, source);
         List<ParameterConfig> parameters = new ArrayList<>();
+        ModelBuilder modelBuilder = new GradleModelBuilder();
         for (PropertyElement property: source.getBeanProperties()) {
-            TypeDef type = ModelUtils.getType(context, element.getPackageName() + ".model",
-                property.getType(), generatedModels);
+            TypeDef type = modelBuilder.getType(context, element.getPackageName() + ".model",
+                property.getType());
             parameters.add(PluginUtils.getParameterConfig(javadoc, property, type));
         }
 
@@ -110,8 +109,8 @@ public final class GradlePluginUtils {
             annotation.stringValue("extensionMethodName").orElse(methodName),
             javadoc.javadoc().orElse(namePrefix + " Gradle task."),
             methodJavadoc,
-            generatedModels,
-            annotation.booleanValue("cacheable").orElse(true)
+            annotation.booleanValue("cacheable").orElse(true),
+            modelBuilder
         );
     }
 
@@ -147,8 +146,8 @@ public final class GradlePluginUtils {
      * @param extensionMethodName The method name for gradle extension
      * @param methodJavadoc The javadoc for executable method
      * @param taskJavadoc The javadoc for the whole task
-     * @param generatedModels Additional generated models
      * @param cacheable Whether the task should be cacheable
+     * @param modelBuilder The model builder
      */
     public record GradleTaskConfig (
         @NonNull ClassElement source,
@@ -158,8 +157,8 @@ public final class GradlePluginUtils {
         @NonNull String extensionMethodName,
         @NonNull String taskJavadoc,
         @NonNull String methodJavadoc,
-        @NonNull List<GeneratedModel> generatedModels,
-        boolean cacheable
+        boolean cacheable,
+        ModelBuilder modelBuilder
     ) {
     }
