+++ micronaut-build-plugin-sourcegen-generator/src/main/java/io/micronaut/sourcegen/generator/visitors/gradle/builder/GradleTaskBuilder.java
@@ -24,8 +24,7 @@ import io.micronaut.inject.ast.ClassElement;
 import io.micronaut.sourcegen.annotations.GenerateGradlePlugin;
 import io.micronaut.sourcegen.annotations.GenerateGradlePlugin.Type;
 import io.micronaut.sourcegen.annotations.PluginTaskParameter.OutputType;
-import io.micronaut.sourcegen.generator.visitors.ModelUtils;
-import io.micronaut.sourcegen.generator.visitors.ModelUtils.GeneratedModel;
+import io.micronaut.sourcegen.generator.visitors.ModelBuilder.GeneratedModel;
 import io.micronaut.sourcegen.generator.visitors.PluginUtils;
 import io.micronaut.sourcegen.generator.visitors.PluginUtils.ParameterConfig;
 import io.micronaut.sourcegen.generator.visitors.gradle.GradlePluginUtils.GradlePluginConfig;
@@ -130,13 +129,15 @@ public class GradleTaskBuilder implements GradleTypeBuilder {
         return List.of(builder.build());
     }
 
-    private MethodDef createParameterGetter(ParameterConfig parameter) {
+    public static MethodDef createParameterGetter(ParameterConfig parameter) {
         MethodDefBuilder propBuilder = MethodDef
             .builder("get" + NameUtils.capitalize(parameter.source().getName()))
             .addModifiers(Modifier.PUBLIC, Modifier.ABSTRACT)
             .addJavadoc(parameter.javadoc())
             .returns(createGradleProperty(parameter));
-        if (parameter.output() != OutputType.NONE) {
+        if (parameter.isPOJO()) {
+            propBuilder.addAnnotation("org.gradle.api.tasks.Nested");
+        } else if (parameter.output() != OutputType.NONE) {
             if (parameter.source().getType().isAssignable(File.class)) {
                 if (parameter.directory()) {
                     propBuilder.addAnnotation(AnnotationDef.builder(ClassTypeDef.of("org.gradle.api.tasks.OutputDirectory")).build());
@@ -191,6 +192,11 @@ public class GradleTaskBuilder implements GradleTypeBuilder {
                         String getterName = "get" + NameUtils.capitalize(parameter.source().getName());
                         TypeDef getterType = createGradleProperty(parameter);
                         ExpressionDef def = t.field(taskField).invoke(getterName, getterType);
+                        if (parameter.isPOJO()) {
+                            statements.add(def.invoke("copyTo", TypeDef.VOID,
+                                params.get(0).invoke(getterName, getterType)));
+                            continue;
+                        }
                         if (!parameter.required()) {
                             if (parameter.defaultValue() != null) {
                                 TypeDef type = parameter.type();
@@ -264,6 +270,9 @@ public class GradleTaskBuilder implements GradleTypeBuilder {
                 .builder("get" + NameUtils.capitalize(parameter.source().getName()))
                 .addModifiers(Modifier.PUBLIC, Modifier.ABSTRACT)
                 .returns(createGradleProperty(parameter));
+            if (parameter.isPOJO()) {
+                propBuilder.addAnnotation("org.gradle.api.tasks.Nested");
+            }
             builder.addMethod(propBuilder.build());
         }
         return builder.build();
@@ -283,7 +292,7 @@ public class GradleTaskBuilder implements GradleTypeBuilder {
                 parametersType
             ))
             .addJavadoc("The work action that actually runs the task logic.")
-            .addMethods(taskConfig.generatedModels().stream().map(GeneratedModel::convertorMethod).toList())
+            .addMethods(taskConfig.modelBuilder().getGeneratedModels().stream().map(GeneratedModel::convertorMethod).toList())
             .addModifiers(Modifier.PUBLIC, Modifier.ABSTRACT, Modifier.STATIC)
             .addMethod(executeMethod)
             .build();
@@ -321,14 +330,16 @@ public class GradleTaskBuilder implements GradleTypeBuilder {
                     TypeDef.parameterized(ClassTypeDef.of("org.gradle.api.provider.Provider"), File.class)
                 );
             }
-            if (!parameter.required() && parameter.defaultValue() == null) {
-                expression = expression.invoke("getOrNull", parameter.type());
-            } else {
-                expression = expression.invoke("get", parameter.type());
+            if (!parameter.isPOJO()) {
+                if (!parameter.required() && parameter.defaultValue() == null) {
+                    expression = expression.invoke("getOrNull", parameter.type());
+                } else {
+                    expression = expression.invoke("get", parameter.type());
+                }
             }
             params.put(
                 parameter.source().getName(),
-                ModelUtils.convertParameterIfRequired(
+                taskConfig.modelBuilder().convertParameterIfRequired(
                     parameter.source().getType(), parameter.source().getName() + "Param", statements, expression
                 )
             );
@@ -337,7 +348,7 @@ public class GradleTaskBuilder implements GradleTypeBuilder {
         return StatementDef.multi(statements);
     }
 
-    static TypeDef createGradleProperty(ParameterConfig parameter) {
+    public static TypeDef createGradleProperty(ParameterConfig parameter) {
         ClassElement type = parameter.source().getType();
         if (type.isAssignable(File.class)) {
             if (parameter.directory()) {
@@ -363,6 +374,9 @@ public class GradleTaskBuilder implements GradleTypeBuilder {
                     parameterized.typeArguments().get(0)
                 );
             }
+        }
+        if (parameter.isPOJO()) {
+            return parameter.type();
         }
         return TypeDef.parameterized(
             ClassTypeDef.of("org.gradle.api.provider.Property"),
